refactor: reorganize test file structure
- Move test files to appropriate directories
- Improve test organization and discoverability
- Add visibility annotations for test-only members
- Improve test isolation and maintainability
