+++ lib/extensions/state_observable_extensions.dart
@@ -1,0 +1,40 @@
+import 'dart:async';
+
+import 'package:flutter/material.dart';
+import 'package:flutter_deep_state_manage/controllers/state_observable.dart';
+
+@visibleForTesting
+extension ObservableStream<T> on StateObservable<T> {
+  @visibleForTesting
+  Stream<T> asStream() {
+    StreamController<T> streamController = StreamController<T>();
+
+    streamController.add(state); //InitialState
+
+    void callback() {
+      streamController.add(state); //LoadingState -> SuccessState
+    }
+
+    addListener(callback);
+
+    return streamController.stream;
+  }
+}
+
+@visibleForTesting
+extension ObservableValueNotifier<T> on ValueNotifier<T> {
+  @visibleForTesting
+  Stream<T> asStream() {
+    StreamController<T> streamController = StreamController<T>();
+
+    streamController.add(value);
+
+    void callback() {
+      streamController.add(value);
+    }
+
+    addListener(callback);
+
+    return streamController.stream;
+  }
+}
