+++ test/testable/controllers/observable_counter.dart
@@ -1,0 +1,13 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_deep_state_manage/controllers/change_state.dart';
+
+@visibleForTesting
+class ObservableCounter extends ChangeState {
+  int counter = 0;
+
+  @visibleForTesting
+  void increment() {
+    counter++;
+    notifyCallback();
+  }
+}
