+++ test/testable/entities/product_entity.dart
@@ -1,0 +1,9 @@
+import 'package:flutter/material.dart';
+
+@visibleForTesting
+class Product {
+  final String id;
+  final String name;
+
+  Product({required this.id, required this.name});
+}
