+++ test/testable/states/base_state.dart
@@ -1,0 +1,20 @@
+import 'package:flutter/material.dart';
+
+@visibleForTesting
+abstract class BaseState {}
+
+class InitialState extends BaseState {}
+
+class LoadingState extends BaseState {}
+
+class SuccessState<T extends Object> extends BaseState {
+  final T data;
+
+  SuccessState({required this.data});
+}
+
+class ErrorState extends BaseState {
+  final String message;
+
+  ErrorState({required this.message});
+}
